% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffDiff_byY.R
\name{aggreg_DID}
\alias{aggreg_DID}
\title{Aggreg results}
\usage{
aggreg_DID(x, DiD_keep = c(1, 4), by_DiD = FALSE, ...)
}
\arguments{
\item{x}{output from DD}

\item{DiD_keep}{which to look at}

\item{by_DiD}{DiD specific?}

\item{\ldots}{grouping vars... time}
}
\description{
Aggreg results
}
