% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiD_manual.R
\name{DD_manu}
\alias{DD_manu}
\title{Diff Diff manual table}
\usage{
DD_manu(
  data,
  y_var = "y",
  time.index = "Time",
  treat = "tr",
  unit.index = "unit",
  control_gr = "0_0",
  treat_gr = "0_1",
  time_subset = NULL,
  subperiod = NULL
)
}
\arguments{
\item{data}{A data frame containing the variables of interest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{y_var}{The name of the y variable}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{control_gr}{The control group of interest}

\item{treat_gr}{The treatment group of interest}

\item{time_subset}{The restriction of time period (if gave more than 2 periods)}

\item{subperiod}{A variable indicating how to group, exclude time periods. See details.}
}
\description{
Diff Diff manual table
}
\examples{
dat_DiD_1 <- sim_dat_staggered(Time=2, timing_treatment=2)
DD_manu(data=dat_DiD_1)

## Compare with felm
library(lfe)
felm(y~tr|unit+Time, data=dat_DiD_1)

## DiD for longer sequences
dat_DiD_long <- sim_dat_staggered(Time=10, timing_treatment=6)
DD_manu(data=dat_DiD_long,  control_gr=paste(rep(0, 10), collapse = "_"),
        treat_gr=paste(c(rep(0, 5), rep(1,5)), collapse = "_"),
        subperiod= rep(c("pre", "post"), each=5))
DD_manu(data=dat_DiD_long,  control_gr=paste(rep(0, 10), collapse = "_"),
        treat_gr=paste(c(rep(0, 5), rep(1,5)), collapse = "_"),
        subperiod= rep(c("placebo", "pre", "post", "dynamic"), c(4,1,1,4)))
felm(y~tr|unit+Time, data=dat_DiD_long)

}
