% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffDiff_manual_byY.R
\name{DD_manu_many}
\alias{DD_manu_many}
\title{Multi-year diff and diff, manually}
\usage{
DD_manu_many(
  y_var = "y",
  data,
  time.index = "Time",
  treat = "tr",
  unit.index = "unit",
  lag = 1,
  format_long = TRUE
)
}
\arguments{
\item{y_var}{The name of the y variable}

\item{data}{A data frame containing the variables of interest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{lag}{How many lags to look at? Lag=1 will consider 0_1 sequences, lag 2: 0_1_1 etc}

\item{format_long}{Should it format in long or wide?}
}
\description{
This function computes the year-to-year Did. This is useful mainly for visualisation, in
particular for the pre-trends (settng lag>1), but has no inference for the DiD effect itself.
}
