% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffDiff_byY.R
\name{DiD_count}
\alias{DiD_count}
\alias{DiD_aggreg}
\title{Count the treated/control}
\usage{
DiD_count(x)

DiD_aggreg(x, DiD_keep = c(1, 4), by_DiD = FALSE, ...)
}
\arguments{
\item{x}{output from DD}

\item{DiD_keep}{which to look at}

\item{by_DiD}{DiD specific?}

\item{\ldots}{grouping vars... time}
}
\description{
Count the treated/control

Aggreg results
}
\examples{
data <- sim_dat(N=100)
DD_out <- DD(data=data)
DiD_aggreg(x=DD_out, by_DiD = FALSE)
DiD_aggreg(x=DD_out, by_DiD = TRUE)
DiD_count(x=DD_out)
}
