% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEdecompo_simple.R
\name{FE_decompo}
\alias{FE_decompo}
\title{Decomposition of fixed-effects weights}
\usage{
FE_decompo(
  data,
  y_var = "y",
  time.index = "Time",
  treat = "tr",
  unit.index = "unit",
  by = c("time", "unit", "both")
)
}
\arguments{
\item{data}{A data frame containing the variables of itnerest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{y_var}{The name of the y variable}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{by}{over which dimension to do it?}
}
\description{
Decomposition of fixed-effects weights
}
\examples{
library(lfe)
 dat_sim_1 <- sim_dat(N=100, Time = 5)
coef_FE_time <- coef(felm(y~tr|Time, data = dat_sim_1))
coef_FE_unit <- coef(felm(y~tr|unit, data = dat_sim_1))

coefs_by_Y <- FE_decompo(data=dat_sim_1,
treat = "tr",
                      by = "time")
coefs_by_U <- FE_decompo(data=dat_sim_1,
              treat = "tr",
                      by = "unit")
with(coefs_by_Y, weighted.mean(treat_coef, treat_weight))
coef_FE_time
with(coefs_by_U, weighted.mean(treat_coef, treat_weight))
coef_FE_unit

}
