% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docu.R
\docType{data}
\name{GentzkowData}
\alias{GentzkowData}
\title{Dataset from Gentzkwow et al (2011)}
\format{
A data-frame
}
\usage{
GentzkowData
}
\description{
This is the replication dataset put by Reed (2019)
}
\examples{
library(multiDiff)
data(GentzkowData)
GentzkowData

## The FE is
}
\references{
Reed, W. Robert, 2019, "Replication Data for: "GUEST BLOG at TRN: EIR - Heterogeneity in Two-Way Fixed Effects Models"", https://doi.org/10.7910/DVN/EGMLQG, Harvard Dataverse, V2, UNF:6:r9g/aJRpVkv2rxw3F3J7QQ==
}
\keyword{datasets}
