% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_CS.R
\name{mdd_CS}
\alias{mdd_CS}
\title{Apply CS estimate}
\usage{
mdd_CS(mdd_dat, ...)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{...}{further objects passed to \code{\link[did]{att_gt}}}
}
\description{
Apply CS estimate
}
\examples{
# example code
dat <- sim_dat_staggered(as_mdd = TRUE, Time=6)
mdd_CS(dat)
}
