% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{mdd_data_format}
\alias{mdd_data_format}
\alias{print.mdd_dat}
\alias{plot.mdd_dat}
\title{Diff Diff manual table}
\usage{
mdd_data_format(
  data,
  y_var = "y",
  time.index = "Time",
  treat = "tr",
  unit.index = "unit"
)

\method{print}{mdd_dat}(x, ...)

\method{plot}{mdd_dat}(x, conf.int = FALSE, ...)
}
\arguments{
\item{data}{A data frame containing the variables of interest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{y_var}{The name of the y variable}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{x}{object of class mdd_dat}

\item{...}{unused}

\item{conf.int}{Whether the plot should show confidence intervals to,
see \code{\link{mdd_group_means}}}
}
\description{
Diff Diff manual table
}
\examples{
dat_DiD_raw <- sim_dat_common()
dat_DiD <- mdd_data_format(dat_DiD_raw)
}
