% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd_DD_means.R
\name{mdd_DD_means22}
\alias{mdd_DD_means22}
\title{Estimate DiD 2x2 means}
\usage{
mdd_DD_means22(mdd_dat, add_tests = TRUE)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{add_tests}{compute also tests?}
}
\description{
Estimate DiD 2x2 means
}
\examples{
dat_sim <- sim_dat_common(timing_treatment = 6:10)
dat_sim_mdd <- mdd_data_format(dat_sim)
out <- mdd_DD_means22(mdd_dat=dat_sim_mdd)
}
