% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_ChaiseFeuille_2020.R
\name{mdd_estim_CH_2020}
\alias{mdd_estim_CH_2020}
\alias{mdd_estim_CH}
\title{de Chaisemartin and d'Haufeuille estimators}
\usage{
mdd_estim_CH_2020(mdd_dat, ...)

mdd_estim_CH(mdd_dat, mode = c("dyn", "had", "old"), ...)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{...}{Further arguments passed to \link[DIDmultiplegt]{did_multiplegt}}

\item{mode}{type of estimator, see \link[DIDmultiplegt]{did_multiplegt}.}
}
\description{
Simple wrapper for  \link[DIDmultiplegt]{did_multiplegt}, see help file there
}
\examples{
data(GentzkowData)
GentzkowData_md <- mdd_data_format(GentzkowData,  y_var="prestout",
                                   time.index = "year", treat = "numdailies", unit.index="cnty90" )
CH_2024_out <- mdd_estim_CH(mdd_dat = GentzkowData_md)
CH_2024_out
tidy(CH_2024_out)
}
\seealso{
\code{\link{mDid_weights_CH}} for a local implementation of the CH weights.
}
