% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_study.R
\name{mdd_event_study}
\alias{mdd_event_study}
\alias{plot.mdd_event_study}
\title{Event study}
\usage{
mdd_event_study(
  data,
  y_var = "y",
  time.index = "Time",
  treat = "tr",
  unit.index = "unit",
  trim_low = NULL,
  trim_high = NULL,
  time.omit = -1,
  weights = NULL
)

\method{plot}{mdd_event_study}(x, ...)
}
\arguments{
\item{data}{A data frame containing the variables of itnerest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{y_var}{The name of the y variable}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{trim_low, trim_high}{Upper/lower bound on parameters to include}

\item{time.omit}{Which is the base year omitted in the analysis?}

\item{weights}{Variable containing the weights}

\item{x}{the ES object}

\item{...}{currently not used}
}
\description{
Event study
}
