% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsynth_diff.R
\name{mdd_gsynth}
\alias{mdd_gsynth}
\title{Run generalized synthetic diff-diff from XXX}
\usage{
mdd_gsynth(mdd_dat, ...)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{...}{Further argumetns passed to}
}
\description{
Run generalized synthetic diff-diff from XXX
}
\examples{
if(require(gsynth)){
 data(gsynth)
 mdd_simdata_gs <- mdd_data_format(simdata,
                                   y_var = "Y",time.index = "time",
                                   treat = "D", unit.index = "id")
 res <- mdd_gsynth(mdd_dat=mdd_simdata_gs)
}
}
