% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo_test.R
\name{mdd_test_placebo}
\alias{mdd_test_placebo}
\title{Run placebo test for DiD}
\usage{
mdd_test_placebo(mdd_dat, T_treated = NULL, weights = NULL, cluster = NULL)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{T_treated}{Number of pre-period to consider as treated}

\item{cluster, weights}{see \code{\link{mdd_DD_simple}}}
}
\description{
Subset the data to the pre-period, and assign a treatment value of 1 to units
eventually treated
}
\examples{
DID_dat_raw <- sim_dat_common(timing_treatment = 5:10)
DID_dat <- mdd_data_format(DID_dat_raw)
mdd_test_placebo(mdd_dat = DID_dat, T_treated = 1)
mdd_test_placebo(mdd_dat = DID_dat, T_treated = 2)
}
