% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_CH.R
\name{mDid_weights_CH}
\alias{mDid_weights_CH}
\alias{print.FE_weights_CH}
\alias{plot.FE_weights_CH}
\alias{weights_CH_smry}
\title{Weights from Chaisemartin and Hautefeuille}
\usage{
mDid_weights_CH(
  data,
  y_var = "y",
  time.index = "Time",
  treat = "tr",
  unit.index = "unit"
)

\method{print}{FE_weights_CH}(x, ...)

\method{plot}{FE_weights_CH}(x, by = c("year", "unit"), ...)

weights_CH_smry(data)
}
\arguments{
\item{data}{A data frame containing the variables of itnerest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{y_var}{The name of the y variable}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{x}{Output from \code{mDid_weights_CH}}

\item{by}{Which kind of plot to draw}

\item{\ldots}{Unused arguments}
}
\description{
Weights from Chaisemartin and Hautefeuille
}
\examples{
examp_df <- data.frame(unit = rep(c(1, 2), each=3),
                      time = rep(1:3, 2),
                      treat = c(0, 0, 1, 0, 1, 1))
mDid_weights_CH(examp_df, treat = "treat",
               time.index = "time")
}
