% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDiff-package.R
\docType{package}
\name{multiDiff}
\alias{multiDiff}
\title{multiDiff: multi-period diff-and-diff}
\description{
This R package implements the multi-period diff-diff as discussed in
Imai and Kim (2019) and Chaisemartin and Hautfeuille (2019),
as well as the Gibbons et al decomposition, and simple event study function.
}
\details{
Main functions are \code{\link{DD}}, and \code{\link{DiD_aggreg}}.
Function \code{\link{sim_dat}} simulates a dataset for the example.

To install:
devtools::install_github("MatthieuStigler/multiDiff", upgrade = "never")
}
