% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pre_trend.R
\name{test_pre_trend}
\alias{test_pre_trend}
\title{Conduct parallel pre-trends test}
\usage{
test_pre_trend(
  data,
  treat.group.index,
  time.index,
  y_var = "y",
  cluster = NULL
)
}
\arguments{
\item{data}{data-frame, containing only pre-intervention values}

\item{treat.group.index}{the variable in \code{data} containing the post-intervention treatment status}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{y_var}{The name of the y variable}

\item{cluster}{argument passed to \code{feols(..., cluster=)}}
}
\description{
Conduct parallel pre-trends test
}
\examples{

data <- multiDiff::sim_dat_staggered(N=100, perc_always = 0, Time=10,
                                     timing_treatment = 6, perc_treat=0.5)
 library(tidyverse, warn.conflicts = FALSE)
 data_modif <- data \%>\%
   group_by(unit) \%>\%
   mutate(type = if_else(any(tr==1), "Treat", "Control")) \%>\%
   ungroup() \%>\%
   select(y, type, Time, unit)

#
test_pre_trend(data_modif, treat.group.index=type,
               time.index=Time, cluster = "unit")

}
