% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_diff.R
\name{tidy.synthdid_estimate}
\alias{tidy.synthdid_estimate}
\title{Tidy a 'synthdid_estimate' object}
\usage{
\method{tidy}{synthdid_estimate}(x, conf.int = FALSE, conf.level = 0.95, method = "jackknife", ...)
}
\arguments{
\item{x}{output from \code{\link{mdd_synthdid}} or synthdid::synthdid_estimate}

\item{conf.int, conf.level, ...}{as standard}

\item{method}{method used in synthdid::vcov.synthdid_estimate}
}
\description{
Tidy a 'synthdid_estimate' object
}
