% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffDiff_byY.R
\name{DD}
\alias{DD}
\title{Compute Diff diff by period}
\usage{
DD(y_var = "y", data, time.index = "Time", treat = "tr",
  unit.index = "unit")
}
\arguments{
\item{y_var}{The name of the y variable}

\item{data}{data}

\item{time.index}{time.index}

\item{treat}{treatment variable}

\item{unit.index}{unit.index}
}
\description{
Compute Diff diff by period
}
\examples{
data <- sim_dat(N=100)
DD_out <- DD(data=data)
DiD_aggreg(x=DD_out, by_DiD = FALSE)
DiD_aggreg(x=DD_out, by_DiD = TRUE)
}
