% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffDiff_byY.R
\name{DD}
\alias{DD}
\title{Compute the diff-and-diff by period. \lifecycle{experimental}}
\usage{
DD(
  y_var = "y",
  data,
  time.index = "Time",
  treat = "tr",
  unit.index = "unit",
  min_obs_required = 2,
  treatment = c("binary", "categorical")
)
}
\arguments{
\item{y_var}{The name of the y variable}

\item{data}{A data frame containing the variables of interest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{min_obs_required}{the minimum number of either treat or control needed to run a regresion}

\item{treatment}{Is treatment binary or categorical?}
}
\description{
Compute the diff-and-diff by period. \lifecycle{experimental}
}
\examples{
data <- sim_dat(N=100)
DD_out <- DD(data=data)
DiD_aggreg(x=DD_out, by_DiD = FALSE)
DiD_aggreg(x=DD_out, by_DiD = TRUE)
}
\seealso{
\code{\link{DiD_aggreg}} for aggregating the output of \code{DD} over years.
}
