% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiD_manual.R
\name{DD_manu}
\alias{DD_manu}
\title{Diff Diff manual table}
\usage{
DD_manu(
  data,
  y_var = "y",
  time.index = "Time",
  treat = "tr",
  unit.index = "unit",
  control_gr = "0_0",
  treat_gr = "0_1",
  time_per = NULL
)
}
\arguments{
\item{data}{A data frame containing the variables of itnerest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{y_var}{The name of the y variable}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{control_gr}{The control group of interest}

\item{treat_gr}{The treatment group of interest}

\item{time_per}{The restriction of time period (if gave more than 2 periods)}
}
\description{
Diff Diff manual table
}
\examples{
data <- sim_dat(N=100, Time =2)
DD_manu(data=data)
}
