% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lag_group}
\alias{lag_group}
\title{Lag by group}
\usage{
lag_group(df, group_var, time_var, lag_var, lagamount = 1)
}
\arguments{
\item{df}{data frame}

\item{group_var}{group_var}

\item{time_var}{time_var}

\item{lag_var}{lag_var}

\item{lagamount}{lag}
}
\description{
Lag by group
}
\examples{
df_test <- data.frame(group = rep(c("a", "b"), each=6),
year = rep(2000:2005, 2),
value = (0:11) ^ 2)
lag_group(df_test, group, time_var=year, lag_var=value)
}
