% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lag_group}
\alias{lag_group}
\title{Lag a variable by group}
\usage{
lag_group(
  df,
  group_var,
  value_var,
  time_var,
  lagamount = 1,
  complete_time = seq(min(df[[time_var]]), max(df[[time_var]]), by = 1L),
  default = NA
)
}
\arguments{
\item{df}{data frame}

\item{group_var}{The grouping variable}

\item{value_var}{The variable to lag over}

\item{time_var}{The time variable.
This is necessary as the function makes sure there are no missing years.}

\item{lagamount}{The amount of lag}

\item{complete_time}{Sequence to use for 'completing' df, making sure implicit NAs are turned into explicit ones.}

\item{default}{Default value for missing? NA by default.}
}
\description{
Take the lag of a variable for each group.
Will check also that there are no missing years.
}
\examples{
df_test <- data.frame(group = rep(c("a", "b"), each=6),
                      year = rep(2000:2005, 2),
                      value = (0:11) ^ 2,
                      value2 = rnorm(12))
lag_group(df_test, "group", time_var="year", value_var="value", lagamount = -1:2)

## Behaviour with missing years
lag_group(df_test[-4,], "group", time_var="year", value_var="value", lagamount = -1:2)

## Many lags and many variables:

lag_group(df_test[-4,], "group", time_var="year", lagamount = 1:2, value_var = c("value", "value2"))
}
