% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_CS.R
\name{mdd_CS}
\alias{mdd_CS}
\title{Apply Callaway and Sant'Anna (2021) estimate}
\usage{
mdd_CS(mdd_dat, timing_treat_var = NULL, panel = NULL, ...)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{timing_treat_var}{variable indicating the timing to treat. Not necessary, unless the dataset is not balanced}

\item{panel}{Argument \code{panel} to \code{\link[did]{att_gt}}. If not specified, will try to guess.}

\item{...}{further objects passed to \code{\link[did]{att_gt}}}
}
\value{
Output of \code{\link[did]{att_gt}}, therefore of class \code{\link[did]{MP}}.
}
\description{
Simple wrapper to function \code{\link[did]{att_gt}} from package did.
}
\examples{
dat <- sim_dat_staggered(as_mdd = TRUE, Time=6)
mdd_CS(dat)

## with a single treatment time, this is equivalent to mdd_event_study()
dat_common <- sim_dat_common(as_mdd = TRUE, timing_treatment = 5:10)
all.equal(unname(coef(mdd_event_study(dat_common))),
          broom::tidy(mdd_CS(dat_common, base_period = "universal"))$estimate[-4])
}
