% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_study.R
\name{mdd_DD_simple}
\alias{mdd_DD_simple}
\title{Estimate simple did}
\usage{
mdd_DD_simple(
  data,
  y_var = "y",
  time.index = "Time",
  treat = "tr",
  unit.index = "unit"
)
}
\arguments{
\item{data}{A data frame containing the variables of itnerest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{y_var}{The name of the y variable}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}
}
\description{
Estimate simple did
}
