% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_simple.R
\name{mdd_DD_simple}
\alias{mdd_DD_simple}
\title{Estimate simple did}
\usage{
mdd_DD_simple(mdd_dat, weights = NULL, cluster = NULL)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{weights}{Weight argument passed to the underlying \code{feols} function}

\item{cluster}{Cluster argument passed to the underlying \code{feols} function}
}
\description{
Standard two-way FE estimation of a DiD
}
\examples{
## simulate and format data
DID_dat <- sim_dat_common(timing_treatment = 5:10, as_mdd=TRUE)

## Estimate DiD
mdd_DD_simple(DID_dat)

## estimate with weights
# add random unit-year specific weights:
DID_dat$wght <- runif(nrow(DID_dat))
mdd_DD_simple(DID_dat, weights=~wght)
}
\seealso{
\code{\link{mdd_event_study}} for the event study.
}
