% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_study.R
\name{mdd_event_study}
\alias{mdd_event_study}
\alias{plot.mdd_event_study}
\title{Event study}
\usage{
mdd_event_study(
  mdd_dat,
  trim_low = NULL,
  trim_high = NULL,
  time.omit = -1,
  weights = NULL
)

\method{plot}{mdd_event_study}(x, ...)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{trim_low, trim_high}{Upper/lower bound on parameters to include}

\item{time.omit}{Which is the base year omitted in the analysis?}

\item{weights}{Variable containing the weights}

\item{x}{the ES object}

\item{...}{currently not used}
}
\description{
Event study
}
\examples{
## simulate and format data
DID_dat_raw <- sim_dat_common(timing_treatment = 5:10)
DID_dat <- mdd_data_format(DID_dat_raw)

## estimate ES
ES_out <- mdd_event_study(DID_dat)
ES_out
summary(ES_out)
plot(ES_out)
}
\seealso{
\code{\link{mdd_test_pre_trend_event}} to run a parallel trend assumption
}
