% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_gsynth.R
\name{mdd_gsynth}
\alias{mdd_gsynth}
\alias{mdd_estim_fect}
\title{Run Generalized Synthetic Control Method (Xu 2017)}
\usage{
mdd_gsynth(mdd_dat, echo = FALSE, parallel = FALSE, ...)

mdd_estim_fect(mdd_dat, echo = FALSE, parallel = FALSE, ...)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{echo}{Whether to print the messages of gsynth or not. Default to TRUE}

\item{parallel}{Argument passed to \link[gsynth]{gsynth}, default is FALSE.}

\item{...}{Further arguments passed to \link[gsynth]{gsynth}}
}
\description{
Simple wrapper for  \link[gsynth]{gsynth}, see help file there
}
\details{
Note that package gsynth is not imported, only suggested.
}
\examples{
if(require(gsynth)){
 data(gsynth)
 mdd_simdata_gs <- mdd_data_format(simdata,
                                   y_var = "Y",time.index = "time",
                                   treat = "D", unit.index = "id")
 res <- mdd_gsynth(mdd_dat=mdd_simdata_gs)
}
}
\seealso{
\link{mdd_synthdid} for the synthetic diff-diff by Arkhangelsky et al (2019).
}
