% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_diff.R
\name{mdd_synthdid}
\alias{mdd_synthdid}
\title{Run synthetic diff-diff}
\usage{
mdd_synthdid(mdd_dat, add_weights = FALSE)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{add_weights}{return the weights?}
}
\description{
Run synthetic diff-diff
}
\examples{
if(require(synthdid)){
  data('california_prop99')
  mdd_california_prop99 <- mdd_data_format(california_prop99,
                                           y_var = "PacksPerCapita",time.index = "Year",
                                           treat = "treated", unit.index = "State")
  mdd_synthdid(mdd_dat=mdd_california_prop99)
}
}
