% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pre_trend.R
\name{mdd_test_pre_trend_means}
\alias{mdd_test_pre_trend_means}
\alias{mdd_test_pre_trend_event}
\title{Conduct parallel pre-trends test}
\usage{
mdd_test_pre_trend_means(mdd_dat, cluster = NULL, time_ref = NULL)

mdd_test_pre_trend_event(mdd_dat, vcov_fix = FALSE, ...)
}
\arguments{
\item{mdd_dat}{An object of class mdd_dat created by \code{\link{mdd_data_format}}}

\item{cluster}{argument passed to \code{feols(..., cluster=)}}

\item{time_ref}{the reference time period against which to test, By default, the first pre-period}

\item{vcov_fix}{argument passed to vcov, in case it is applied to a fixest object}

\item{...}{Further arguments passed to \code{\link{mdd_event_study}}}
}
\description{
Conduct parallel pre-trends test
}
\examples{

data <- sim_dat_common(N=100, Time=10,
                       timing_treatment = 6:10, perc_treat=0.5)

mdd_data <- mdd_data_format(data)
mdd_test_pre_trend_means(mdd_dat=mdd_data)

}
\seealso{
\code{\link{mdd_event_study}} to estimate the event study.
}
