% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mDid_weights_CH.R
\name{mDid_weights_CH}
\alias{mDid_weights_CH}
\alias{print.FE_weights_CH}
\alias{plot.FE_weights_CH}
\alias{weights_CH_smry}
\title{Weights from Chaisemartin and Hautefeuille}
\usage{
mDid_weights_CH(
  data,
  y_var = "y",
  time.index = "Time",
  treat = "tr",
  unit.index = "unit",
  return_details = FALSE
)

\method{print}{FE_weights_CH}(x, ...)

\method{plot}{FE_weights_CH}(x, by = c("time", "unit"), ...)

weights_CH_smry(data)
}
\arguments{
\item{data}{A data frame containing the variables of itnerest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{y_var}{The name of the y variable}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{return_details}{Return additional details on weights?}

\item{x}{Output from \code{mDid_weights_CH}}

\item{by}{Which kind of plot to draw}

\item{\ldots}{Unused arguments}
}
\value{
A enhanced \code{tibble} of class \emph{FE_weights_CH}
\describe{
\item{w}{The weight \eqn{w_{g,t}} as defined in the paper}
\item{weight}{The weight w multiplied by \code{obs_weight}. They sum to 1.}
\item{obs_weight}{The ratio \eqn{N_{g,t}/N_1}, eventually including also the treatment intensity.}
}
}
\description{
Weights from Chaisemartin and Hautefeuille
}
\examples{
examp_df <- data.frame(unit = rep(c(1, 2), each=3),
                      time = rep(1:3, 2),
                      treat = c(0, 0, 1, 0, 1, 1))
mDid_weights_CH(examp_df, treat = "treat",
               time.index = "time")
}
