% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mDid_weights_CH.R
\name{mDid_weights_CH}
\alias{mDid_weights_CH}
\alias{print.FE_weights_CH}
\alias{plot.FE_weights_CH}
\alias{weights_CH_smry}
\title{Weights from Chaisemartin and Hautefeuille (2020)}
\usage{
mDid_weights_CH(
  data,
  y_var = "y",
  time.index = "Time",
  treat = "tr",
  unit.index = "unit",
  return_details = FALSE
)

\method{print}{FE_weights_CH}(x, ...)

\method{plot}{FE_weights_CH}(x, by = c("time", "unit"), ...)

weights_CH_smry(data)
}
\arguments{
\item{data}{A data frame containing the variables of interest, should have names  \code{y_var}, \code{treat}, \code{time.index} and
\code{unit.index}}

\item{y_var}{The name of the y variable}

\item{time.index}{The name of the variable describing the time index (date, year, etc)}

\item{treat}{The name of the treatment variable.}

\item{unit.index}{The name of the units (individual, county, household, etc).}

\item{return_details}{Return additional details on weights?}

\item{x}{Output from \code{mDid_weights_CH}}

\item{by}{Which kind of plot to draw}

\item{\ldots}{Unused arguments}
}
\value{
A enhanced \code{tibble} of class \emph{FE_weights_CH}
\describe{
\item{w}{The weight \eqn{w_{g,t}} as defined in the paper}
\item{weight}{The weight w multiplied by \code{obs_weight}. They sum to 1.}
\item{obs_weight}{The ratio \eqn{N_{g,t}/N_1}, eventually including also the treatment intensity.}
}
}
\description{
Weights from Chaisemartin and Hautefeuille (2020)
}
\examples{
## example from page 2970 in Chaisemartin and Hautefeuille (2020)
examp_df <- data.frame(unit = rep(c(1, 2), each=3),
                      Time = rep(1:3, 2),
                      treat = c(0, 0, 1, 0, 1, 1))
w_df <- mDid_weights_CH(examp_df, treat = "treat",
                        time.index = "Time")
w_df
## the three weights shown in the beta^FE equation:
subset(as.data.frame(w_df), treat==1, "weight")

## replicate application B of Gentzkow, Shapiro, and Sinkinson (2011)
data(GentzkowData)
data_W_CH <- mDid_weights_CH(data=GentzkowData,
                             y_var="prestout",
                             time.index = "year",
                             treat = "numdailies",
                             unit.index="cnty90")
## results are a bit different compared to the paper:
data_W_CH
plot(data_W_CH, by = "unit")
plot(data_W_CH, by = "time")
}
\references{
de Chaisemartin and D'Haultfoeuille (2020), Two-Way Fixed Effects Estimators with Heterogeneous Treatment Effects
American Economic Review , Vol. 110, No. 9
}
\seealso{
\code{\link{mdd_estim_CH}} for the CH estimators and \code{\link{GentzkowData}} for a presentation of the Gentzkow datasets
}
