% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDiff.R
\docType{package}
\name{multiDiff}
\alias{multiDiff}
\alias{multiDiff-package}
\title{multiDiff: Internal functions I use}
\description{
to do... can I get a list of functions? Just document them!
}
\details{
Should run:
remotes::install_github("MatthieuStigler/multiDiff", upgrade = "never")
}
