% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_sim.R
\name{sim_dat}
\alias{sim_dat}
\title{\subsection{NEW}{

SImulate
}}
\usage{
sim_dat(
  N = 1000,
  Time = 15,
  beta = 1,
  gamma = 0.7,
  seed = NULL,
  prob_treat = 0.25
)
}
\arguments{
\item{N}{number of distinct units}

\item{Time}{number of distinct time}

\item{beta}{coef}

\item{gamma}{To generate the series \code{y_asym}, which has longer dynamics.
Gamma is the coefficients for observations that were not treated (0) the previous period,
while the standard beta will apply for values currently treated.}

\item{seed}{seed}

\item{prob_treat}{Probability of treatment}
}
\description{
\subsection{NEW}{

SImulate
}
}
\examples{
sim_dat()
}
